refactor(smooth-movement): split implementation into focused files
