+++ docs/plugins/smooth-movement.rst
@@ -18,7 +18,6 @@ Usage
 
     enable smooth-movement
     smooth-movement
-    smooth-movement camera on|off|reset|<x-offset> <y-offset>
     smooth-movement flip on|off
     smooth-movement zlevel on|off
 
@@ -35,27 +34,11 @@ Examples
 ``smooth-movement flip on``
     Horizontally flip creatures so they face their direction of travel.
 
-``smooth-movement camera on``
-    Enable smooth camera scrolling and pixel-precise middle-mouse dragging.
-
-``smooth-movement camera 0.25 -0.25``
-    Enable the free camera and offset the view by a quarter tile east and
-    north of the tile grid.
-
-``smooth-movement camera reset``
-    Return the free camera to the tile grid without disabling it.
-
 ``smooth-movement zlevel on``
     Also animate movement on visible z-levels below the main level.
 
 Settings
 --------
-
-``camera`` (default: off)
-    Enable or disable smooth camera scrolling and free-camera dragging. The
-    ``reset`` argument clears the persistent sub-tile offset. Two numeric
-    arguments set the horizontal and vertical offsets in tiles; each must be
-    between -0.99 and 0.99.
 
 ``flip`` (default: off)
     Flip creature sprites horizontally according to their last horizontal
