+++ plugins/smooth-movement/visual_renderer.cpp
@@ -1,0 +1,890 @@
+// SPDX-License-Identifier: MIT
+
+#include <algorithm>
+#include <array>
+#include <cmath>
+#include <cstdint>
+#include <set>
+#include <string>
+#include <type_traits>
+#include <unordered_map>
+#include <utility>
+#include <vector>
+
+#include <SDL_render.h>
+
+#include "Core.h"
+#include "MemAccess.h"
+#include "PluginManager.h"
+
+#include "modules/DFSDL.h"
+
+#include "df/graphic.h"
+#include "df/graphic_viewportst.h"
+#include "df/renderer_2d_base.h"
+#include "df/texture_fullid.h"
+
+#include "visual_animation.h"
+#include "visual_renderer.h"
+
+using namespace DFHack;
+
+using df::global::gps;
+using df::global::window_x;
+using df::global::window_y;
+using df::global::window_z;
+
+namespace {
+
+// Runtime harness for the engine-owned visual state; gameplay data is never
+// read.
+decltype(&SDL_RenderCopyF) render_copy_f = nullptr;
+decltype(&SDL_RenderCopyExF) render_copy_ex_f = nullptr;
+decltype(&SDL_RenderFillRect) render_fill_rect = nullptr;
+decltype(&SDL_GetRenderDrawColor) get_render_draw_color = nullptr;
+decltype(&SDL_SetRenderDrawColor) set_render_draw_color = nullptr;
+
+visual_animation_managerst animation_manager;
+std::set<std::pair<int32_t, int32_t>> previous_coverage;
+uint64_t visual_context_revision = 0;
+const void *previous_viewport = nullptr;
+std::array<int32_t, 12> previous_view_signature{};
+bool has_view_signature = false;
+// Map scroll (window_x/window_y) is tracked separately from the reset
+// signature: a pure pan is followed (movements are translated) instead of
+// triggering a full reset, so it must NOT bump the context revision. It only
+// invalidates the viewport-space blackout coverage from the prior frame.
+int32_t previous_pan_x = 0;
+int32_t previous_pan_y = 0;
+bool has_pan_context = false;
+bool flip_enabled = false;
+bool zlevel_enabled = false;
+
+constexpr uint32_t fire_bits = 0x70000000U;
+
+void update_visual_context(const df::renderer_2d_base *renderer, const df::graphic_viewportst *vp) {
+    // window_x/window_y are deliberately excluded: a horizontal/vertical scroll
+    // is followed, not reset. window_z (z-level) stays, since a z change is not
+    // followable.
+    const std::array<int32_t, 12> signature = {window_z ? *window_z : 0,
+                                               vp->dim_x,
+                                               vp->dim_y,
+                                               vp->clipx[0],
+                                               vp->clipx[1],
+                                               vp->clipy[0],
+                                               vp->clipy[1],
+                                               renderer->viewport_zoom_factor,
+                                               renderer->origin_x,
+                                               renderer->origin_y,
+                                               gps->dimx,
+                                               gps->dimy};
+    const bool changed =
+        !has_view_signature || previous_viewport != vp || previous_view_signature != signature;
+    if (changed) {
+        ++visual_context_revision;
+        previous_coverage.clear();
+    }
+    previous_viewport = vp;
+    previous_view_signature = signature;
+    has_view_signature = true;
+
+    // On a pure pan the reset signature is unchanged, but last frame's blackout
+    // coverage is in the old viewport frame, so discard it (the engine repaints
+    // the whole scrolled viewport anyway).
+    const int32_t pan_x = window_x ? *window_x : 0;
+    const int32_t pan_y = window_y ? *window_y : 0;
+    if (!has_pan_context || previous_pan_x != pan_x || previous_pan_y != pan_y)
+        previous_coverage.clear();
+    previous_pan_x = pan_x;
+    previous_pan_y = pan_y;
+    has_pan_context = true;
+}
+
+using viewport_layer_memberst = int32_t *df::graphic_viewportst::*;
+
+struct visual_layer_bufferst {
+    viewport_visual_layer layer;
+    viewport_layer_memberst current;
+    viewport_layer_memberst previous;
+};
+
+constexpr size_t visual_layer_count = static_cast<size_t>(viewport_visual_layer::count);
+constexpr std::array visual_layer_buffers = {
+    visual_layer_bufferst{viewport_visual_layer::right,
+                          &df::graphic_viewportst::screentexpos_right_creature,
+                          &df::graphic_viewportst::screentexpos_right_creature_old},
+    visual_layer_bufferst{viewport_visual_layer::center, &df::graphic_viewportst::screentexpos,
+                          &df::graphic_viewportst::screentexpos_old},
+    visual_layer_bufferst{viewport_visual_layer::left,
+                          &df::graphic_viewportst::screentexpos_left_creature,
+                          &df::graphic_viewportst::screentexpos_left_creature_old},
+    visual_layer_bufferst{viewport_visual_layer::upright,
+                          &df::graphic_viewportst::screentexpos_upright_creature,
+                          &df::graphic_viewportst::screentexpos_upright_creature_old},
+    visual_layer_bufferst{viewport_visual_layer::up,
+                          &df::graphic_viewportst::screentexpos_up_creature,
+                          &df::graphic_viewportst::screentexpos_up_creature_old},
+    visual_layer_bufferst{viewport_visual_layer::upleft,
+                          &df::graphic_viewportst::screentexpos_upleft_creature,
+                          &df::graphic_viewportst::screentexpos_upleft_creature_old},
+    visual_layer_bufferst{viewport_visual_layer::vehicle,
+                          &df::graphic_viewportst::screentexpos_vehicle,
+                          &df::graphic_viewportst::screentexpos_vehicle_old},
+    visual_layer_bufferst{viewport_visual_layer::item, &df::graphic_viewportst::screentexpos_item,
+                          &df::graphic_viewportst::screentexpos_item_old},
+    visual_layer_bufferst{viewport_visual_layer::designation,
+                          &df::graphic_viewportst::screentexpos_designation,
+                          &df::graphic_viewportst::screentexpos_designation_old}};
+
+constexpr bool valid_visual_layer_buffers() {
+    uint16_t layers = 0;
+    for (const auto &buffer : visual_layer_buffers) {
+        const uint16_t layer = uint16_t(1U << static_cast<uint8_t>(buffer.layer));
+        if (layers & layer)
+            return false;
+        layers |= layer;
+    }
+    return layers == uint16_t((1U << visual_layer_count) - 1);
+}
+
+static_assert(valid_visual_layer_buffers());
+
+template <typename Viewport> auto visual_layers(Viewport *vp, bool previous = false) {
+    using layer_pointer = std::conditional_t<std::is_const_v<Viewport>, const int32_t *, int32_t *>;
+    std::array<layer_pointer, visual_layer_count> layers{};
+    for (const auto &buffer : visual_layer_buffers)
+        layers[static_cast<size_t>(buffer.layer)] =
+            vp->*(previous ? buffer.previous : buffer.current);
+    return layers;
+}
+
+viewport_visual_animation_inputst animation_input(df::graphic_viewportst *vp) {
+    const df::graphic_viewportst *const_viewport = vp;
+    return {vp,
+            vp->dim_x,
+            vp->dim_y,
+            visual_context_revision,
+            visual_layers(const_viewport),
+            visual_layers(const_viewport, true),
+            window_x ? *window_x : 0,
+            window_y ? *window_y : 0};
+}
+
+// The layer buffers are freed and nulled without clearing the active flag.
+bool viewport_readable(df::graphic_viewportst *vp) {
+    return vp != nullptr && vp->flag.bits.active && animation_input(vp).valid();
+}
+
+int32_t tile_pixel(int32_t tile, int32_t origin, int32_t zoom) {
+    return zoom == 128 ? 32 * tile + origin : (zoom * 32 * tile) / 128 + origin;
+}
+
+bool inside_clip(const df::graphic_viewportst *vp, int32_t x, int32_t y) {
+    return x >= vp->clipx[0] && x <= vp->clipx[1] && y >= vp->clipy[0] && y <= vp->clipy[1];
+}
+
+bool has_fire(const df::graphic_viewportst *vp, int32_t x, int32_t y) {
+    return vp->screentexpos_spatter_flag != nullptr &&
+           (vp->screentexpos_spatter_flag[x * vp->dim_y + y] & fire_bits) != 0;
+}
+
+template <typename T> class scoped_value_restorest {
+    T &value;
+    T saved;
+
+  public:
+    explicit scoped_value_restorest(T &value, T replacement = T{})
+        : value(value), saved(std::exchange(value, std::move(replacement))) {
+        static_assert(std::is_nothrow_move_assignable_v<T>);
+    }
+
+    ~scoped_value_restorest() noexcept { value = std::move(saved); }
+
+    scoped_value_restorest(const scoped_value_restorest &) = delete;
+    scoped_value_restorest &operator=(const scoped_value_restorest &) = delete;
+    scoped_value_restorest(scoped_value_restorest &&) = delete;
+    scoped_value_restorest &operator=(scoped_value_restorest &&) = delete;
+};
+
+template <typename Callback> void with_zeroed_values(const Callback &callback) { callback(); }
+
+template <typename Callback, typename T, typename... Values>
+void with_zeroed_values(const Callback &callback, T &value, Values &...values) {
+    scoped_value_restorest<T> zero(value);
+    with_zeroed_values(callback, values...);
+}
+
+struct render_proxyst {
+    viewport_visual_layer layer;
+    float source_x;
+    float source_y;
+    int32_t target_x;
+    int32_t target_y;
+    int32_t texpos;
+    float progress;
+    SDL_Texture *texture;
+    bool mirrored = false;
+    int32_t mirror_shift = 0;
+    std::set<std::pair<int32_t, int32_t>> coverage;
+};
+
+using tile_coveragest = std::set<std::pair<int32_t, int32_t>>;
+
+struct render_coveragest {
+    tile_coveragest all;
+    std::array<tile_coveragest, static_cast<size_t>(visual_render_groupst::count)> groups;
+    std::unordered_map<int32_t, uint16_t> selected;
+};
+
+struct viewport_renderst {
+    df::graphic_viewportst *viewport;
+    std::vector<render_proxyst> proxies;
+    render_coveragest coverage;
+};
+
+constexpr uint16_t visual_layer_bit(viewport_visual_layer layer) {
+    return uint16_t(1U << static_cast<uint8_t>(layer));
+}
+
+uint16_t selected_mask(const std::unordered_map<int32_t, uint16_t> &selected, int32_t index) {
+    const auto found = selected.find(index);
+    return found == selected.end() ? 0 : found->second;
+}
+
+template <size_t Layer = 0, typename Callback>
+void with_suppressed_visual_layers(const std::array<int32_t *, visual_layer_count> &layers,
+                                   int32_t index, uint16_t mask, const Callback &callback) {
+    if constexpr (Layer == visual_layer_count)
+        callback();
+    else if (mask & (1U << Layer)) {
+        scoped_value_restorest<int32_t> zero(layers[Layer][index]);
+        with_suppressed_visual_layers<Layer + 1>(layers, index, mask, callback);
+    } else
+        with_suppressed_visual_layers<Layer + 1>(layers, index, mask, callback);
+}
+
+template <typename Callback>
+void with_base_suppressed(df::graphic_viewportst *vp, int32_t index, const Callback &callback) {
+    with_zeroed_values(callback, vp->screentexpos_background[index],
+                       vp->screentexpos_floor_flag[index], vp->screentexpos_background_two[index],
+                       vp->screentexpos_liquid_flag[index], vp->screentexpos_spatter_flag[index],
+                       vp->screentexpos_spatter[index], vp->screentexpos_ramp_flag[index],
+                       vp->screentexpos_shadow_flag[index], vp->screentexpos_building_one[index]);
+}
+
+template <typename Callback>
+void with_main_suppressed(df::graphic_viewportst *vp, int32_t index, const Callback &callback) {
+    with_base_suppressed(vp, index,
+                         [&] { with_zeroed_values(callback, vp->screentexpos_vermin[index]); });
+}
+
+template <typename Callback>
+void with_upper_suppressed(df::graphic_viewportst *vp, int32_t index, const Callback &callback) {
+    with_main_suppressed(vp, index, [&] {
+        with_zeroed_values(callback, vp->screentexpos_building_two[index],
+                           vp->screentexpos_projectile[index], vp->screentexpos_high_flow[index],
+                           vp->screentexpos_top_shadow[index], vp->screentexpos_signpost[index]);
+    });
+}
+
+void redraw_viewport_tile(df::renderer_2d_base *renderer, const viewport_renderst &viewport,
+                          int32_t x, int32_t y, bool defer_interface) {
+    df::graphic_viewportst *vp = viewport.viewport;
+    const int32_t index = x * vp->dim_y + y;
+    const auto redraw = [&] { renderer->update_viewport_tile(vp, x, y); };
+    const auto stage = [&] {
+        with_suppressed_visual_layers(visual_layers(vp), index,
+                                      selected_mask(viewport.coverage.selected, index), redraw);
+    };
+    // The interface layer is the shading for levels below the camera.
+    // A staged tile has a sprite drawn over it afterwards, so draw_interface_only
+    // places it instead.
+    if (!defer_interface || vp->screentexpos_interface == nullptr)
+        stage();
+    else
+        with_zeroed_values(stage, vp->screentexpos_interface[index]);
+}
+
+// Every buffer the interface-only pass zeroes has to exist before it can be
+// zeroed.
+bool interface_pass_readable(const df::graphic_viewportst *vp) {
+    return vp != nullptr && vp->screentexpos_interface != nullptr &&
+           vp->screentexpos_background != nullptr && vp->screentexpos_floor_flag != nullptr &&
+           vp->screentexpos_background_two != nullptr && vp->screentexpos_liquid_flag != nullptr &&
+           vp->screentexpos_spatter_flag != nullptr && vp->screentexpos_spatter != nullptr &&
+           vp->screentexpos_ramp_flag != nullptr && vp->screentexpos_shadow_flag != nullptr &&
+           vp->screentexpos_building_one != nullptr && vp->screentexpos_vermin != nullptr &&
+           vp->screentexpos_building_two != nullptr && vp->screentexpos_projectile != nullptr &&
+           vp->screentexpos_high_flow != nullptr && vp->screentexpos_signpost != nullptr;
+}
+
+// Runs after the proxies so the shading covers them rather than sitting
+// underneath.
+void draw_interface_only(df::renderer_2d_base *renderer, df::graphic_viewportst *vp, int32_t x,
+                         int32_t y) {
+    if (!interface_pass_readable(vp))
+        return;
+    const int32_t index = x * vp->dim_y + y;
+    const auto redraw = [&] { renderer->update_viewport_tile(vp, x, y); };
+    const auto without_visuals = [&] {
+        with_suppressed_visual_layers(visual_layers(vp), index,
+                                      uint16_t((1U << visual_layer_count) - 1), redraw);
+    };
+    with_zeroed_values(without_visuals, vp->screentexpos_background[index],
+                       vp->screentexpos_floor_flag[index], vp->screentexpos_background_two[index],
+                       vp->screentexpos_liquid_flag[index], vp->screentexpos_spatter_flag[index],
+                       vp->screentexpos_spatter[index], vp->screentexpos_ramp_flag[index],
+                       vp->screentexpos_shadow_flag[index], vp->screentexpos_building_one[index],
+                       vp->screentexpos_vermin[index], vp->screentexpos_building_two[index],
+                       vp->screentexpos_projectile[index], vp->screentexpos_high_flow[index],
+                       vp->screentexpos_signpost[index]);
+}
+
+void redraw_world_tile(df::renderer_2d_base *renderer,
+                       const std::vector<viewport_renderst> &viewports,
+                       const tile_coveragest &staged, int32_t x, int32_t y) {
+    // The stage pass repaints everything above the lowest across the staged
+    // tiles, after the proxies.
+    const bool staged_tile = staged.count({x, y}) != 0;
+    for (const viewport_renderst &viewport : viewports) {
+        if (inside_clip(viewport.viewport, x, y))
+            redraw_viewport_tile(renderer, viewport, x, y, staged_tile);
+        if (staged_tile)
+            break;
+    }
+}
+
+constexpr uint16_t visual_layers_through_group(visual_render_groupst group) {
+    uint16_t mask = 0;
+    for (const auto &descriptor : visual_layer_descriptors)
+        if (descriptor.render_group != visual_render_groupst::designation &&
+            static_cast<uint8_t>(descriptor.render_group) <= static_cast<uint8_t>(group))
+            mask |= visual_layer_bit(descriptor.layer);
+    return mask;
+}
+
+void redraw_above(df::renderer_2d_base *renderer, df::graphic_viewportst *vp, int32_t x, int32_t y,
+                  visual_render_groupst group,
+                  const std::unordered_map<int32_t, uint16_t> &selected) {
+    const int32_t index = x * vp->dim_y + y;
+    const auto redraw = [&] { renderer->update_viewport_tile(vp, x, y); };
+    const auto suppress_visuals = [&] {
+        const auto stage = [&] {
+            with_suppressed_visual_layers(
+                visual_layers(vp), index,
+                selected_mask(selected, index) | visual_layers_through_group(group), redraw);
+        };
+        // The interface layer sits above every group, so each group's redraw would
+        // paint it again. draw_interface_only places it once, after the sprites.
+        if (vp->screentexpos_interface == nullptr)
+            stage();
+        else
+            with_zeroed_values(stage, vp->screentexpos_interface[index]);
+    };
+    if (group == visual_render_groupst::item || group == visual_render_groupst::vehicle)
+        with_base_suppressed(vp, index, suppress_visuals);
+    else if (group == visual_render_groupst::main)
+        with_main_suppressed(vp, index, suppress_visuals);
+    else
+        with_upper_suppressed(vp, index, suppress_visuals);
+}
+
+SDL_Texture *cached_texture(df::renderer_2d_base *renderer, int32_t texpos,
+                            bool transparent_background = true) {
+    if (texpos == 0)
+        return nullptr;
+    df::texture_fullid texture_id;
+    texture_id.texpos = texpos;
+    texture_id.r = texture_id.g = texture_id.b = 1.0f;
+    texture_id.br = texture_id.bg = texture_id.bb = 0.0f;
+    texture_id.flag =
+        transparent_background ? df::texture_fullid_flag::mask_transparent_background : 0;
+    const auto texture = renderer->tile_cache.tile_cache.find(texture_id);
+    return texture == renderer->tile_cache.tile_cache.end()
+               ? nullptr
+               : static_cast<SDL_Texture *>(texture->second);
+}
+
+// The render_copy_ex_f null check is defensive only, not a graceful-degradation
+// path. `bind` aborts load_sdl on any missing symbol and plugin_enable then
+// refuses the render hook.
+void render_copy_maybe_mirrored(SDL_Renderer *renderer, SDL_Texture *texture,
+                                const SDL_FRect &destination, bool mirrored) {
+    if (mirrored && render_copy_ex_f != nullptr) {
+        render_copy_ex_f(renderer, texture, nullptr, &destination, 0.0, nullptr,
+                         SDL_FLIP_HORIZONTAL);
+        return;
+    }
+    render_copy_f(renderer, texture, nullptr, &destination);
+}
+
+void draw_proxy(df::renderer_2d_base *renderer, const render_proxyst &proxy) {
+    const int32_t zoom = renderer->viewport_zoom_factor;
+    const int32_t target_x = tile_pixel(proxy.target_x, renderer->origin_x, zoom);
+    const int32_t target_y = tile_pixel(proxy.target_y, renderer->origin_y, zoom);
+    const float tile_size = float(zoom == 128 ? 32 : std::max(1, zoom * 32 / 128));
+    const float source_x = target_x + (proxy.source_x - proxy.target_x) * tile_size;
+    const float source_y = target_y + (proxy.source_y - proxy.target_y) * tile_size;
+    const float mirror_offset = float(proxy.mirror_shift) * tile_size;
+    const SDL_FRect destination = {
+        source_x + (target_x - source_x) * proxy.progress + mirror_offset,
+        source_y + (target_y - source_y) * proxy.progress, tile_size, tile_size};
+    render_copy_maybe_mirrored(static_cast<SDL_Renderer *>(renderer->sdl_renderer), proxy.texture,
+                               destination, proxy.mirrored);
+}
+
+std::vector<render_proxyst> collect_proxies(df::renderer_2d_base *renderer,
+                                            df::graphic_viewportst *vp, bool movement_enabled) {
+    std::vector<render_proxyst> proxies;
+    auto layers = visual_layers(vp);
+    auto previous_layers = visual_layers(vp, true);
+    for (uint8_t draw_order = 0; draw_order < visual_layer_count; ++draw_order) {
+        const viewport_visual_layer visual_layer = visual_layer_at_draw_order(draw_order);
+        const size_t layer = static_cast<size_t>(visual_layer);
+        for (int32_t y = 0; y < vp->dim_y; ++y) {
+            for (int32_t x = 0; x < vp->dim_x; ++x) {
+                const int32_t index = x * vp->dim_y + y;
+                const int32_t texpos = layers[layer][index];
+                if (texpos == 0)
+                    continue;
+                if (!movement_enabled)
+                    continue;
+                const auto movement = animation_manager.get_movement(
+                    vp, static_cast<viewport_visual_layer>(layer), x, y);
+                if (!movement.active)
+                    continue;
+                const int32_t inherited_source_x =
+                    inherited_visual_source_tile(x, movement.source_x, x);
+                const int32_t inherited_source_y =
+                    inherited_visual_source_tile(y, movement.source_y, y);
+                const bool inherited_source_in_bounds =
+                    inherited_source_x >= 0 && inherited_source_x < vp->dim_x &&
+                    inherited_source_y >= 0 && inherited_source_y < vp->dim_y;
+                if (!visual_layer_moves_independently(visual_layer)) {
+                    bool anchored = false;
+                    for (const render_proxyst &anchor : proxies) {
+                        if (anchor.layer == viewport_visual_layer::center &&
+                            std::abs(anchor.target_x - x) <= 1 &&
+                            std::abs(anchor.target_y - y) <= 1 &&
+                            anchor.source_x - anchor.target_x == movement.source_x - x &&
+                            anchor.source_y - anchor.target_y == movement.source_y - y &&
+                            anchor.progress == movement.progress)
+                            anchored = true;
+                    }
+                    if (!anchored)
+                        continue;
+                }
+                if ((visual_layer == viewport_visual_layer::item ||
+                     visual_layer == viewport_visual_layer::designation) &&
+                    movement.inherited) {
+                    if (visual_layer == viewport_visual_layer::item &&
+                        vp->screentexpos_old[index] != 0)
+                        continue;
+                    if (!inherited_source_in_bounds)
+                        continue;
+                    const int32_t source = inherited_source_x * vp->dim_y + inherited_source_y;
+                    if (!visual_moved_between_tiles(visual_layer, layers[layer],
+                                                    previous_layers[layer], source, index))
+                        continue;
+                }
+                if (!visual_layer_moves_independently(visual_layer) &&
+                    visual_layer != viewport_visual_layer::designation && movement.inherited) {
+                    const bool fragment_moved =
+                        inherited_source_in_bounds &&
+                        visual_moved_between_tiles(
+                            visual_layer, layers[layer], previous_layers[layer],
+                            inherited_source_x * vp->dim_y + inherited_source_y, index);
+                    if (!fragment_moved) {
+                        const auto &descriptor = visual_layer_descriptor(visual_layer);
+                        bool owns_fragment = false;
+                        for (const render_proxyst &anchor : proxies)
+                            if (anchor.layer == viewport_visual_layer::center &&
+                                anchor.target_x == x + descriptor.center_x &&
+                                anchor.target_y == y + descriptor.center_y &&
+                                anchor.source_x - anchor.target_x == movement.source_x - x &&
+                                anchor.source_y - anchor.target_y == movement.source_y - y &&
+                                anchor.progress == movement.progress)
+                                owns_fragment = true;
+                        if (!owns_fragment)
+                            continue;
+                    }
+                }
+
+                // Items, vehicles and designations keep their vanilla orientation.
+                const auto &mirror_descriptor = visual_layer_descriptor(visual_layer);
+                const visual_render_groupst group = visual_render_group(visual_layer);
+                const bool mirror_eligible =
+                    flip_enabled &&
+                    (group == visual_render_groupst::main || group == visual_render_groupst::upper);
+                // Facing is read from the anchor tile so every fragment of one creature
+                // agrees.
+                const bool mirrored =
+                    mirror_eligible && animation_manager.get_facing(
+                                           vp, x + mirror_descriptor.center_x,
+                                           y + mirror_descriptor.center_y) != native_sprite_facing;
+                // The anchor's own layer has center_x 0, so it flips in place.
+                const int32_t mirror_shift =
+                    mirrored ? mirrored_tile_x(x, x + mirror_descriptor.center_x) - x : 0;
+                render_proxyst proxy = {static_cast<viewport_visual_layer>(layer),
+                                        movement.source_x,
+                                        movement.source_y,
+                                        x,
+                                        y,
+                                        texpos,
+                                        movement.progress,
+                                        nullptr,
+                                        mirrored,
+                                        mirror_shift,
+                                        {}};
+                bool blocked = false;
+                for (int32_t coverage_x = int32_t(std::floor(std::min(proxy.source_x, float(x))));
+                     coverage_x <= int32_t(std::ceil(std::max(proxy.source_x, float(x))));
+                     ++coverage_x) {
+                    for (int32_t coverage_y =
+                             int32_t(std::floor(std::min(proxy.source_y, float(y))));
+                         coverage_y <= int32_t(std::ceil(std::max(proxy.source_y, float(y))));
+                         ++coverage_y) {
+                        if (!inside_clip(vp, coverage_x, coverage_y)) {
+                            blocked = true;
+                            break;
+                        }
+                        if (visual_render_group(proxy.layer) == visual_render_groupst::main &&
+                            has_fire(vp, coverage_x, coverage_y)) {
+                            blocked = true;
+                            break;
+                        }
+                        proxy.coverage.emplace(coverage_x, coverage_y);
+                    }
+                    if (blocked)
+                        break;
+                }
+                if (blocked)
+                    continue;
+                if (proxy.mirror_shift != 0) {
+                    std::set<std::pair<int32_t, int32_t>> mirrored_coverage;
+                    for (const auto &tile : proxy.coverage)
+                        mirrored_coverage.emplace(tile.first + proxy.mirror_shift, tile.second);
+                    for (const auto &tile : mirrored_coverage) {
+                        if (!inside_clip(vp, tile.first, tile.second)) {
+                            blocked = true;
+                            break;
+                        }
+                        if (visual_render_group(proxy.layer) == visual_render_groupst::main &&
+                            has_fire(vp, tile.first, tile.second)) {
+                            blocked = true;
+                            break;
+                        }
+                        proxy.coverage.insert(tile);
+                    }
+                    if (blocked)
+                        continue;
+                }
+
+                proxy.texture = cached_texture(renderer, texpos);
+                if (proxy.texture == nullptr)
+                    continue;
+                proxies.push_back(std::move(proxy));
+            }
+        }
+    }
+
+    // A creature that has stopped still needs its mirrored sprite painted each
+    // frame. Otherwise the engine repaints it natively and the two orientations
+    // alternate between steps. A fragment's tile is its anchor minus the layer's
+    // centre offset, inverting the moving path.
+    if (flip_enabled) {
+        for (int32_t anchor_x = 0; anchor_x < vp->dim_x; ++anchor_x) {
+            for (int32_t anchor_y = 0; anchor_y < vp->dim_y; ++anchor_y) {
+                if (animation_manager.get_facing(vp, anchor_x, anchor_y) == native_sprite_facing)
+                    continue;
+                for (uint8_t draw_order = 0; draw_order < visual_layer_count; ++draw_order) {
+                    const viewport_visual_layer visual_layer =
+                        visual_layer_at_draw_order(draw_order);
+                    const visual_render_groupst group = visual_render_group(visual_layer);
+                    if (group != visual_render_groupst::main &&
+                        group != visual_render_groupst::upper)
+                        continue;
+                    const auto &descriptor = visual_layer_descriptor(visual_layer);
+                    const int32_t x = anchor_x - descriptor.center_x;
+                    const int32_t y = anchor_y - descriptor.center_y;
+                    if (x < 0 || x >= vp->dim_x || y < 0 || y >= vp->dim_y)
+                        continue;
+                    const size_t layer = static_cast<size_t>(visual_layer);
+                    const int32_t texpos = layers[layer][x * vp->dim_y + y];
+                    if (texpos == 0)
+                        continue;
+                    bool already_drawn = false;
+                    for (const render_proxyst &existing : proxies)
+                        if (existing.layer == visual_layer && existing.target_x == x &&
+                            existing.target_y == y)
+                            already_drawn = true;
+                    if (already_drawn)
+                        continue;
+
+                    // source == target at progress 1.0 draws in place, moved only by
+                    // mirror_shift.
+                    render_proxyst proxy = {visual_layer,
+                                            float(x),
+                                            float(y),
+                                            x,
+                                            y,
+                                            texpos,
+                                            1.0f,
+                                            nullptr,
+                                            true,
+                                            mirrored_tile_x(x, anchor_x) - x,
+                                            {}};
+                    // The sprite lands on x+mirror_shift, so that interval must be
+                    // repaintable. The shift has either sign, so order the interval ends
+                    // first.
+                    const int32_t coverage_first = std::min(x, x + proxy.mirror_shift);
+                    const int32_t coverage_last = std::max(x, x + proxy.mirror_shift);
+                    bool blocked = false;
+                    for (int32_t coverage_x = coverage_first; coverage_x <= coverage_last;
+                         ++coverage_x) {
+                        if (!inside_clip(vp, coverage_x, y) ||
+                            (group == visual_render_groupst::main && has_fire(vp, coverage_x, y))) {
+                            blocked = true;
+                            break;
+                        }
+                        proxy.coverage.emplace(coverage_x, y);
+                    }
+                    if (blocked)
+                        continue;
+
+                    proxy.texture = cached_texture(renderer, texpos);
+                    if (proxy.texture == nullptr)
+                        continue;
+                    proxies.push_back(std::move(proxy));
+                }
+            }
+        }
+    }
+    return proxies;
+}
+
+render_coveragest collect_coverage(const std::vector<render_proxyst> &proxies, int32_t dim_y) {
+    render_coveragest coverage;
+    for (const render_proxyst &proxy : proxies) {
+        coverage.all.insert(proxy.coverage.begin(), proxy.coverage.end());
+        coverage.selected[proxy.target_x * dim_y + proxy.target_y] |= visual_layer_bit(proxy.layer);
+        auto &group = coverage.groups[static_cast<size_t>(visual_render_group(proxy.layer))];
+        group.insert(proxy.coverage.begin(), proxy.coverage.end());
+    }
+    return coverage;
+}
+
+std::vector<df::graphic_viewportst *> active_viewports() {
+    std::vector<df::graphic_viewportst *> viewports;
+    if (gps == nullptr)
+        return viewports;
+    for (int32_t lower = 7; lower >= 0; --lower) {
+        df::graphic_viewportst *vp = gps->lower_viewport[lower];
+        if (viewport_readable(vp))
+            viewports.push_back(vp);
+    }
+    if (viewport_readable(gps->main_viewport))
+        viewports.push_back(gps->main_viewport);
+    return viewports;
+}
+
+std::vector<viewport_renderst>
+collect_viewport_renders(df::renderer_2d_base *renderer,
+                         const std::vector<df::graphic_viewportst *> &viewports) {
+    std::vector<viewport_renderst> renders;
+    renders.reserve(viewports.size());
+    for (df::graphic_viewportst *vp : viewports) {
+        const bool movement_enabled = zlevel_enabled || vp == gps->main_viewport;
+        viewport_renderst render = {vp, collect_proxies(renderer, vp, movement_enabled), {}};
+        render.coverage = collect_coverage(render.proxies, vp->dim_y);
+        renders.push_back(std::move(render));
+    }
+    return renders;
+}
+
+tile_coveragest collect_viewport_coverage(const std::vector<viewport_renderst> &viewports) {
+    tile_coveragest coverage;
+    for (const viewport_renderst &viewport : viewports)
+        coverage.insert(viewport.coverage.all.begin(), viewport.coverage.all.end());
+    return coverage;
+}
+
+void draw_interpolation_stages(df::renderer_2d_base *renderer, df::graphic_viewportst *vp,
+                               const std::vector<render_proxyst> &proxies,
+                               const render_coveragest &coverage) {
+    for (size_t index = 0; index < coverage.groups.size(); ++index) {
+        const auto group = static_cast<visual_render_groupst>(index);
+        for (const render_proxyst &proxy : proxies)
+            if (visual_render_group(proxy.layer) == group)
+                draw_proxy(renderer, proxy);
+        if (group == visual_render_groupst::designation)
+            continue;
+        for (const auto &[x, y] : coverage.groups[index])
+            redraw_above(renderer, vp, x, y, group, coverage.selected);
+    }
+}
+
+void redraw_viewport_tiles(df::renderer_2d_base *renderer, const viewport_renderst &viewport,
+                           const tile_coveragest &coverage) {
+    df::graphic_viewportst *vp = viewport.viewport;
+    for (const auto &[x, y] : coverage) {
+        if (!inside_clip(vp, x, y))
+            continue;
+        redraw_viewport_tile(renderer, viewport, x, y, true);
+    }
+}
+
+void draw_viewport_interpolation_stages(df::renderer_2d_base *renderer,
+                                        const std::vector<viewport_renderst> &viewports,
+                                        const tile_coveragest &coverage) {
+    for (size_t index = 0; index < viewports.size(); ++index) {
+        // A lower z-level's proxy must be covered by the next viewport's fog and
+        // terrain. Reapply that viewport before its own proxies, matching DF's
+        // lower-to-main draw order.
+        if (index > 0)
+            redraw_viewport_tiles(renderer, viewports[index], coverage);
+        const viewport_renderst &viewport = viewports[index];
+        draw_interpolation_stages(renderer, viewport.viewport, viewport.proxies, viewport.coverage);
+        // A viewport shades everything drawn beneath it, so this covers every
+        // staged tile. Restricting it to the tiles this viewport has sprites on
+        // would not deepen with distance.
+        for (const auto &[x, y] : coverage) {
+            if (inside_clip(viewport.viewport, x, y))
+                draw_interface_only(renderer, viewport.viewport, x, y);
+        }
+    }
+}
+
+bool has_mirrored_viewport_facing(const std::vector<df::graphic_viewportst *> &viewports) {
+    for (const df::graphic_viewportst *vp : viewports)
+        if (animation_manager.has_mirrored_facing(vp))
+            return true;
+    return false;
+}
+
+void render_interpolated_world(df::renderer_2d_base *renderer) {
+    df::graphic_viewportst *vp = gps ? gps->main_viewport : nullptr;
+    const std::vector<df::graphic_viewportst *> viewports = active_viewports();
+
+    if (vp != nullptr)
+        update_visual_context(renderer, vp);
+    const uint32_t now_ms = Core::getInstance().p->getTickCount();
+    animation_manager.begin_frame(now_ms);
+    for (df::graphic_viewportst *viewport : viewports)
+        animation_manager.synchronize_viewport(animation_input(viewport));
+    animation_manager.end_frame();
+
+    if (!viewport_readable(vp) || renderer->sdl_renderer == nullptr)
+        return;
+    const bool animation_redraw =
+        zlevel_enabled ? animation_manager.requires_full_redraw()
+                       : animation_manager.has_active_movement(vp) || !previous_coverage.empty();
+    if (!animation_redraw && (!flip_enabled || !has_mirrored_viewport_facing(viewports)))
+        return;
+
+    std::vector<viewport_renderst> viewport_renders = collect_viewport_renders(renderer, viewports);
+    tile_coveragest coverage = collect_viewport_coverage(viewport_renders);
+
+    SDL_Renderer *sdl_renderer = static_cast<SDL_Renderer *>(renderer->sdl_renderer);
+    const int32_t zoom = renderer->viewport_zoom_factor;
+    const int32_t tile_size = zoom == 128 ? 32 : std::max(1, zoom * 32 / 128);
+
+    tile_coveragest redraw_coverage = coverage;
+    redraw_coverage.insert(previous_coverage.begin(), previous_coverage.end());
+    Uint8 old_r = 0, old_g = 0, old_b = 0, old_a = 255;
+    get_render_draw_color(sdl_renderer, &old_r, &old_g, &old_b, &old_a);
+    set_render_draw_color(sdl_renderer, 0, 0, 0, 255);
+    for (const auto &[x, y] : redraw_coverage) {
+        if (!inside_clip(vp, x, y))
+            continue;
+        const SDL_Rect tile_rect = {tile_pixel(x, renderer->origin_x, zoom),
+                                    tile_pixel(y, renderer->origin_y, zoom), tile_size, tile_size};
+        render_fill_rect(sdl_renderer, &tile_rect);
+    }
+    set_render_draw_color(sdl_renderer, old_r, old_g, old_b, old_a);
+
+    for (const auto &[x, y] : redraw_coverage) {
+        if (inside_clip(vp, x, y))
+            redraw_world_tile(renderer, viewport_renders, coverage, x, y);
+    }
+    draw_viewport_interpolation_stages(renderer, viewport_renders, coverage);
+
+    previous_coverage = std::move(coverage);
+}
+
+void clear_sdl_bindings() {
+    render_copy_f = nullptr;
+    render_copy_ex_f = nullptr;
+    render_fill_rect = nullptr;
+    get_render_draw_color = nullptr;
+    set_render_draw_color = nullptr;
+}
+
+bool load_sdl(color_ostream &out) {
+    clear_sdl_bindings();
+    DFLibrary *sdl_handle = DFSDL::obtain_library_handle();
+#define bind(name, target)                                                                         \
+    target = reinterpret_cast<decltype(target)>(LookupPlugin(sdl_handle, #name));                  \
+    if (target == nullptr) {                                                                       \
+        out.printerr("smooth-movement: SDL2 function unavailable: " #name "\n");                   \
+        clear_sdl_bindings();                                                                      \
+        return false;                                                                              \
+    }
+    bind(SDL_RenderCopyF, render_copy_f);
+    bind(SDL_RenderCopyExF, render_copy_ex_f);
+    bind(SDL_RenderFillRect, render_fill_rect);
+    bind(SDL_GetRenderDrawColor, get_render_draw_color);
+    bind(SDL_SetRenderDrawColor, set_render_draw_color);
+#undef bind
+    return true;
+}
+
+void reset_state() {
+    animation_manager = visual_animation_managerst();
+    previous_coverage.clear();
+    visual_context_revision = 0;
+    previous_viewport = nullptr;
+    previous_view_signature = {};
+    has_view_signature = false;
+    previous_pan_x = 0;
+    previous_pan_y = 0;
+    has_pan_context = false;
+    flip_enabled = false;
+    zlevel_enabled = false;
+}
+
+} // namespace
+
+namespace smooth_movement {
+
+bool initialize(color_ostream &out) {
+    reset_state();
+    return load_sdl(out);
+}
+
+void shutdown() {
+    reset_state();
+    clear_sdl_bindings();
+    if (gps != nullptr)
+        ++gps->force_full_display_count;
+}
+
+void render(df::renderer_2d_base *renderer) { render_interpolated_world(renderer); }
+
+bool flip_enabled() { return ::flip_enabled; }
+
+void set_flip_enabled(bool enable) {
+    ::flip_enabled = enable;
+    if (gps != nullptr)
+        ++gps->force_full_display_count;
+}
+
+bool zlevel_enabled() { return ::zlevel_enabled; }
+
+void set_zlevel_enabled(bool enable) {
+    ::zlevel_enabled = enable;
+    if (gps != nullptr)
+        ++gps->force_full_display_count;
+}
+
+} // namespace smooth_movement
